Migrate 'is-hidden' test
